+++ src/renderer/features/network/NetworkList/model/__tests__/active-networks-model.test.ts
