+++ src/renderer/features/network/NetworkList/model/__tests__/inactive-networks-model.test.ts
