+++ src/renderer/features/network/NetworkList/model/__tests__/mocks/networks-mock.ts
