+++ src/renderer/features/network/NetworkList/model/inactive-networks-model.ts
@@ -1,4 +1,4 @@
-import { createEvent, createStore, combine, sample } from 'effector';
+import { createEvent, sample, createStore, combine } from 'effector';
 
 import { networkUtils, networkModel } from '@entities/network';
 import { Chain } from '@shared/core';
@@ -8,7 +8,7 @@ const networksChanged = createEvent<Chain[]>();
 
 const $networks = createStore<Chain[]>([]);
 
-const $activeNetworks = combine(
+const $inactiveNetworks = combine(
   {
     networks: $networks,
     connectionStatuses: networkModel.$connectionStatuses,
@@ -17,14 +17,15 @@ const $activeNetworks = combine(
   ({ networks, connectionStatuses, connections }) => {
     return networksListUtils
       .getExtendedChain(networks, connections, connectionStatuses)
-      .filter((o) => networkUtils.isEnabledConnection(o.connection));
+      .filter((o) => networkUtils.isDisabledConnection(o.connection));
   },
 );
 
 sample({ clock: networksChanged, target: $networks });
 
-export const activeNetworksModel = {
-  $activeNetworks,
+export const inactiveNetworksModel = {
+  $inactiveNetworks,
+
   events: {
     networksChanged,
   },
