+++ src/renderer/features/network/NetworkList/ui/ActiveNetwork.tsx
