+++ src/renderer/features/network/NetworkList/ui/EmptyNetworks.tsx
