+++ src/renderer/features/network/NetworkList/ui/InactiveNetwork.tsx
