+++ src/renderer/features/network/NetworkList/ui/styles.css
